+++ src/com/jpexs/decompiler/flash/gui/taglistview/TagListTreeModel.java
@@ -24,6 +24,8 @@ import com.jpexs.decompiler.flash.gui.tagtree.AbstractTagTreeModel;
 import com.jpexs.decompiler.flash.tags.DefineBinaryDataTag;
 import com.jpexs.decompiler.flash.tags.DefineSpriteTag;
 import com.jpexs.decompiler.flash.tags.base.ASMSourceContainer;
+import com.jpexs.decompiler.flash.tags.base.ButtonAction;
+import com.jpexs.decompiler.flash.tags.base.ButtonTag;
 import com.jpexs.decompiler.flash.timeline.Frame;
 import com.jpexs.decompiler.flash.timeline.Timelined;
 import com.jpexs.decompiler.flash.treeitems.HeaderItem;
@@ -97,9 +99,18 @@ public class TagListTreeModel extends AbstractTagTreeModel {
             return ((Frame) parentNode).allInnerTags.get(index);
         } else if (parentNode instanceof DefineBinaryDataTag) {
             return ((DefineBinaryDataTag) parentNode).innerSwf;
-        } else if (parentNode instanceof ASMSourceContainer) {
+        }
+        
+        if (parentNode instanceof ButtonTag) {
+            if (index < ((ButtonTag) parentNode).getRecords().size()) {
+                return ((ButtonTag) parentNode).getRecords().get(index);
+            }
+            index -= ((ButtonTag) parentNode).getRecords().size();
+        }
+        if (parentNode instanceof ASMSourceContainer) {
             return ((ASMSourceContainer)parentNode).getSubItems().get(index);
         }
+        
         throw new Error("Unsupported parent type: " + parentNode.getClass().getName());
     }
 
@@ -107,7 +118,7 @@ public class TagListTreeModel extends AbstractTagTreeModel {
     public int getChildCount(Object parent) {
         TreeItem parentNode = (TreeItem) parent;
         if (parentNode == root) {
-            return  swfs.size();
+            return swfs.size();
         } else if (parentNode instanceof SWFList) {
             return  ((SWFList) parentNode).swfs.size();
         } else if (parentNode instanceof SWF) {
@@ -120,11 +131,17 @@ public class TagListTreeModel extends AbstractTagTreeModel {
             return  ((DefineSpriteTag) parentNode).getTimeline().getFrameCount();
         } else if (parentNode instanceof DefineBinaryDataTag) {
             return  (((DefineBinaryDataTag) parentNode).innerSwf == null ? 0 : 1);
-        } else if (parentNode instanceof ASMSourceContainer) {
-            return ((ASMSourceContainer)parentNode).getSubItems().size();
         }
+        int size = 0;
+        if (parentNode instanceof ButtonTag) {
+            size += ((ButtonTag) parentNode).getRecords().size();
+        }
+        if (parentNode instanceof ASMSourceContainer) {
+            size += ((ASMSourceContainer)parentNode).getSubItems().size();
+        }        
+        
 
-        return 0;
+        return size;
     }
 
     @Override
@@ -165,9 +182,23 @@ public class TagListTreeModel extends AbstractTagTreeModel {
             return ((Frame) parentNode).allInnerTags.indexOf(child);
         } else if (parentNode instanceof DefineBinaryDataTag) {
             return ((DefineBinaryDataTag) parentNode).innerSwf == child ? 0 : -1;
-        } else if (parentNode instanceof ASMSourceContainer) {
-            return ((ASMSourceContainer)parentNode).getSubItems().indexOf(child);
         }
+        int base = 0;
+        if (parentNode instanceof ButtonTag) {
+            int index = ((ButtonTag) parentNode).getRecords().indexOf(child);
+            if (index > -1) {
+                return index;
+            }
+            base = ((ButtonTag) parentNode).getRecords().size();
+        }
+        if (parentNode instanceof ASMSourceContainer) {
+            int index = ((ASMSourceContainer)parentNode).getSubItems().indexOf(child);
+            if (index == -1) {
+                return -1;
+            }
+            return base + index;
+        }
+        
         return -1;
     }
 
@@ -270,11 +301,16 @@ public class TagListTreeModel extends AbstractTagTreeModel {
             } else {
                 return new ArrayList<>(0);
             }
-        } else if (parentNode instanceof ASMSourceContainer) {
-            return ((ASMSourceContainer)parentNode).getSubItems();
         }
+        List<TreeItem> ret = new ArrayList<>();
+        if (parentNode instanceof ButtonTag) {
+            ret.addAll(((ButtonTag) parentNode).getRecords());
+        }
+        if (parentNode instanceof ASMSourceContainer) {
+            ret.addAll(((ASMSourceContainer)parentNode).getSubItems());
+        }        
 
-        return new ArrayList<>();
+        return ret;
     }
     
     @Override